Fix perl regex to use named backreferences
If you create a project that starts with a number, `DB_DATABASE` is lost in the updated `.env` file. This is because original regex uses `\1` numbered backreference. When project starts with a number, the first digit of the project name is joined with `1` to point to unknown backreference: `\17milos`.

To fix this, I used [named backreferences](https://www.regular-expressions.info/replacebackref.html#named).